backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

# База данных
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
geoalchemy2==0.18.4

# Аутентификация
pyjwt[crypto]==2.8.0
pwdlib[argon2]==0.2.1
python-multipart==0.0.12
pyotp==2.9.0

# Celery + Redis
celery==5.4.0
redis==5.1.1
fakeredis

# Rate Limiting
slowapi==0.1.9

# Monitoring
sentry-sdk[fastapi]==2.18.0

# AI/ML
scikit-learn==1.5.2

# Утилиты
pydantic-settings==2.5.2
python-dotenv==1.0.1
pydantic[email]
jinja2==3.1.4
requests==2.32.3

# Tests
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

shapely==2.0.6
cloudinary==1.40.0
python-multipart==0.0.12   # уже есть — для загрузки файлов
Pillow==10.4.0             # работа с изображениями + EXIF
